compat: fix pointer→int cast for pointer difference in W3DAssetManager
Pointer subtraction (mesh_name - name) produces a correct ptrdiff_t;
casting individual pointer values to int then subtracting loses the
upper 32 bits on ARM64/x86_64. Use (int)(mesh_name - name) instead.

+++ GeneralsMD/Code/GameEngineDevice/Source/W3DDevice/GameClient/W3DAssetManager.cpp
@@ -799,7 +799,7 @@ RenderObjClass * W3DAssetManager::Create_Render_Obj(
 		const char *mesh_name = ::strchr (name, '.');
 		if (mesh_name != NULL) 
 		{
-			::lstrcpyn(filename, name, ((int)mesh_name) - ((int)name) + 1);
+			::lstrcpyn(filename, name, (int)(mesh_name - name) + 1);
 #ifdef	INCLUDE_GRANNY_IN_BUILD
 			if (isGranny)
 				::lstrcat(filename, ".gr2");
